fix(ModalSort): One sort modal
