+++ components/base/Filters/index.tsx
@@ -1,5 +1,3 @@
 export { default as FilterCategories } from './FilterCategories';
 export { default as FilterPrice } from './FilterPrice';
 export { default as FilterTypeSales } from './FilterTypeSales';
-export { default as SortDate } from './SortDate';
-export { default as SortPopularity } from './SortPopularity';
